code fixes and refactor
