+++ src/ui/server/helpers/mappings/map-application-to-form-fields/index.ts
@@ -3,7 +3,7 @@ import { FIELD_IDS } from '../../../constants';
 import formatDate from '../../date/format-date';
 import getDateFieldsFromTimestamp from '../../date/get-date-fields-from-timestamp';
 import mapFinancialYearEndDate from '../map-financial-year-end-date';
-import transformNumberToString from '../../display-number-field-value';
+import transformNumberToString from '../../transform-number-to-string';
 
 const {
   SUBMISSION_DEADLINE,
