+++ src/ui/server/helpers/summary-lists/generate-field-group-item/index.ts
@@ -2,7 +2,7 @@ import getKeyText from '../get-key-text';
 import { DEFAULT } from '../../../content-strings';
 import { objectHasProperty } from '../../object';
 import { SummaryListItemData, SummaryListItemDataInput } from '../../../../types';
-import transformNumberToString from '../../display-number-field-value';
+import transformNumberToString from '../../transform-number-to-string';
 
 /**
  * getSummaryListItemDataValue
@@ -18,18 +18,20 @@ export const getSummaryListItemDataValue = (fieldId: string, data?: object, cust
     return String(customValue);
   }
 
-  if (data && objectHasProperty(data, fieldId)) {
-    return String(data[fieldId]);
-  }
+  if (data) {
+    if (objectHasProperty(data, fieldId)) {
+      return String(data[fieldId]);
+    }
 
-  /**
-   * if data is the number 0, then transform to string
-   * 0 fails the objectHasProperty(data, fieldId) as read as false
-   */
-  if (data && transformNumberToString(data[fieldId])) {
-    const value = transformNumberToString(data[fieldId]);
+    /**
+     * if data is the number 0, then transform to string
+     * 0 fails the objectHasProperty(data, fieldId) as read as false
+     */
+    if (transformNumberToString(data[fieldId])) {
+      const value = transformNumberToString(data[fieldId]);
 
-    return String(value);
+      return String(value);
+    }
   }
 
   return DEFAULT.EMPTY;
